+++ docroot/themes/contrib/civictheme/civictheme_library/components/00-base/_variables.components.scss
@@ -288,10 +288,10 @@ $civictheme-iframe-dark-background-color: civictheme-color('primary', 5) !defaul
 //
 // Field label.
 //
-$civictheme-field-label-light-color: civictheme-color('neutral', 7) !default;
-$civictheme-field-label-light-required-color: civictheme-color('error') !default;
-$civictheme-field-label-dark-color: civictheme-color('neutral', 1) !default;
-$civictheme-field-label-dark-required-color: civictheme-color('error') !default;
+$civictheme-label-light-color: civictheme-color('neutral', 7) !default;
+$civictheme-label-light-required-color: civictheme-color('error') !default;
+$civictheme-label-dark-color: civictheme-color('neutral', 1) !default;
+$civictheme-label-dark-required-color: civictheme-color('error') !default;
 
 //
 // Table.
