+++ docroot/themes/contrib/civictheme/civictheme_library/components/02-molecules/pagination/pagination.twig
@@ -29,7 +29,7 @@
 
     {% if items_per_page %}
       <div class="civictheme-pager__items_per_page">
-        {%- include '@atoms/field-label/field-label.twig' with {
+        {%- include '@atoms/label/label.twig' with {
           content: items_per_page_label
         } -%}
         {%- include '@atoms/select/select.twig' with {
