+++ docroot/themes/contrib/civictheme/civictheme_library/components/03-organisms/password-indicator/password-indicator.stories.twig
@@ -26,7 +26,7 @@
   'data-password-indicator-setting-min-count="12"',
 ]|join(' ') %}
 {% set username_label %}
-  {% include '@atoms/field-label/field-label.twig' with {
+  {% include '@atoms/label/label.twig' with {
     theme: theme,
     content: 'Username',
     attributes: 'for="username"',
@@ -40,7 +40,7 @@
   } only %}
 {% endset %}
 {% set password_label %}
-  {% include '@atoms/field-label/field-label.twig' with {
+  {% include '@atoms/label/label.twig' with {
     theme: theme,
     content: 'Password',
     attributes: 'for="password"',
@@ -54,7 +54,7 @@
   } only %}
 {% endset %}
 {% set confirm_label %}
-  {% include '@atoms/field-label/field-label.twig' with {
+  {% include '@atoms/label/label.twig' with {
     theme: theme,
     title: 'Confirm Password',
     attributes: 'for="confirm"',
