+++ docroot/themes/contrib/civictheme/templates/form/form-element-label.html.twig
@@ -5,4 +5,4 @@
  */
 #}
 
-{% include '@atoms/field-label/field-label.twig' %}
+{% include '@atoms/label/label.twig' %}
